feat: implement CPU and CUDA job template plugins with preparation and job creation hooks
