+++ q8s/plugins/job_template_spec.py
@@ -1,7 +1,7 @@
 import os
 from typing import Dict
 import pluggy
-from kubernetes import client, config
+from kubernetes import client
 
 from q8s.constants import WORKSPACE
 from q8s.enums import Target
@@ -13,6 +13,19 @@ hookimpl = pluggy.HookimplMarker("q8s")
 
 
 class JobTemplatePluginSpec:
+
+    @hookspec
+    def prepare(
+        self,
+        target: Target,
+        name: str,
+        namespace: str,
+        env: Dict[
+            str,
+            str | None,
+        ],
+    ) -> None:
+        pass
 
     @hookspec
     def makejob(
@@ -29,141 +42,6 @@ class JobTemplatePluginSpec:
     ) -> client.V1PodTemplateSpec:
         return None
 
-
-class CPUJobTemplatePlugin:
-
-    @hookimpl
-    def makejob(
-        self,
-        name: str,
-        registry_pat: str | None,
-        registry_credentials_secret_name: str,
-        container_image: str,
-        env: Dict[
-            str,
-            str | None,
-        ],
-        target: Target,
-    ) -> client.V1PodTemplateSpec:
-
-        if target != Target.cpu:
-            return None
-
-        container = client.V1Container(
-            name="quantum-routine",
-            image=container_image,
-            env=env,
-            command=["python"],
-            args=[f"{WORKSPACE}/main.py"],
-            volume_mounts=[
-                client.V1VolumeMount(
-                    name="app-volume", mount_path=WORKSPACE, read_only=True
-                )
-            ],
-        )
-
-        template = client.V1PodTemplateSpec(
-            metadata=client.V1ObjectMeta(labels={"app": name}),
-            spec=client.V1PodSpec(
-                containers=[container],
-                image_pull_secrets=(
-                    [
-                        client.V1LocalObjectReference(
-                            name=registry_credentials_secret_name
-                        )
-                    ]
-                    if registry_pat
-                    else []
-                ),
-                restart_policy="Never",
-                volumes=[
-                    client.V1Volume(
-                        name="app-volume",
-                        config_map=client.V1ConfigMapVolumeSource(name=name),
-                    )
-                ],
-            ),
-        )
-
-        return template
-
-
-class CUDAJobTemplatePlugin:
-
-    @hookimpl
-    def makejob(
-        self,
-        name: str,
-        registry_pat: str | None,
-        registry_credentials_secret_name: str,
-        container_image: str,
-        env: Dict[
-            str,
-            str | None,
-        ],
-        target: Target,
-    ) -> client.V1PodTemplateSpec:
-
-        if target != Target.gpu:
-            return None
-
-        container = client.V1Container(
-            name="quantum-routine",
-            image=container_image,
-            env=env,
-            command=["python"],
-            args=[f"{WORKSPACE}/main.py"],
-            resources=(
-                client.V1ResourceRequirements(
-                    limits=(
-                        {
-                            "cpu": "2",
-                            "ephemeral-storage": "50Gi",
-                            "memory": MEMORY,
-                            "nvidia.com/gpu": "1",
-                            # "qubernetes.dev/qpu": "1",
-                        }
-                    ),
-                    requests=(
-                        {
-                            "cpu": "2",
-                            "ephemeral-storage": "0",
-                            "memory": MEMORY,
-                            "nvidia.com/gpu": "1",
-                            # "qubernetes.dev/qpu": "1",
-                        }
-                    ),
-                )
-            ),
-            volume_mounts=[
-                client.V1VolumeMount(
-                    name="app-volume", mount_path=WORKSPACE, read_only=True
-                )
-            ],
-        )
-
-        template = client.V1PodTemplateSpec(
-            metadata=client.V1ObjectMeta(labels={"app": name}),
-            spec=client.V1PodSpec(
-                containers=[container],
-                image_pull_secrets=(
-                    [
-                        client.V1LocalObjectReference(
-                            name=registry_credentials_secret_name
-                        )
-                    ]
-                    if registry_pat
-                    else []
-                ),
-                runtime_class_name="nvidia",
-                restart_policy="Never",
-                volumes=[
-                    client.V1Volume(
-                        name="app-volume",
-                        config_map=client.V1ConfigMapVolumeSource(name=name),
-                    )
-                ],
-            ),
-        )
-
-        return template
+    @hookspec
+    def cleanup(self, name: str, namespace: str) -> None:
+        pass
